Fix race condition with response attribute. Fixes #12.
There are two use cases as I see it. One is to fire off a series of requests
and react to all of them. The other is to treat all responses but the response to the most recent request as stale (e.g. in a search app you only want to show results for the current search and ignore out of order responses)

This change accomodates both use cases by firing events for every response (
which is unchanged), but setting up the response attribute to be either the
response for the most resent request, or null. Likewise it exposes an error
attribute that behaves the same way.