Convert all tests to use promises.
Also improve our tests for errors. Ran into a difficult bug that
was the result of a combination of window.onerror firing because
iron-ajax's error event had bubbled up, and the fact that we were
using webcomponents.js, which wrapped the CustomEvent in some
cases, which interfered with event.stopPropagation() on Firefox.