Request timeouts are errors.
This was trickier than it looks, as the XHR object doesn't look any different
after it has timed out, and the readystatechange handler is called before
the timeout handler is called.

Happily, we can move readystatechange to loadend, which is called *after*
all other events. We then also keep track of whether the request has timed
out ourselves.