Move from readystatechanged to loadend.
readystatechanged is fired *before* any other end-of-request events
are called. This is sub-optimal because we have to guess whether the
request is successful or it failed based on not a lot of information.

If we instead put this guessing code in loadend which fires *after*,
and use the 'error' event to catch network errors, we can catch more
kinds of error that aren't reflected in properties on the XHR object..

This PR also changes our manual JSON parse behavior to return null
if we can't parse the JSON. This is so that our json XHR polyfill
behaves more like the native behavior. i.e. we should only parse the
JSON if the browser didn't do it for us, and if the browser does it for
us, it's spec'd to return null if there's a parse error.