No automatic requests without a url.
We've got this bug where we send unintended requests when the URL hasn't yet been set: https://github.com/PolymerElements/iron-ajax/issues/55

The fix that people have been proposing is that we not fire requests when the URL is the empty string. The tricky thing there is that an empty string URL is totally valid, it refers to the current URL. This isn't just a theoritical use either, it's not uncommon for servers to return different results for a URL based on method, params, accept headers, etc.

This commit tries to thread the needle of back-compat and user intent. The URL now defaults to `undefined`, and we won't perform any automatic requests while the url is `null` or `undefined`. If you explictly set the url to `''` then requests will fire, and if you call generateRequest with a falsey url we'll treat it as `''`.

This will hopefully solve the unintended request problem while also giving users control of when requests will and won't be fired (by setting the url to `null`), while having minimal impact on users that are happy with the current behavior.

Fixes #55