Fixing rare 'indexOf' of undefined error
When a user sets parameters but not a url, iron-ajax will crash (Cannot read property 'indexOf' of undefined). This happens in the get requestUrl() function.

One could fix it there by adding an if() statement in the requestUrl getter, but then we fire unwanted requests again (issue #55).
Therefore I would propose adding an if (!this.url) statement to generateRequest(). It will once and for all squash #55.